% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tested.R
\name{tested}
\alias{tested}
\title{Demo function to return numbers}
\usage{
tested(x)
}
\arguments{
\item{x}{character or numeric vector}
}
\value{
numeric vector
}
\description{
Demo function to return numbers
}

